Fixed spacing between notification windows.
The fix was to modernize the return type in the declaration of -requiredDistanceFromExistingDisplays in GrowlPositionController's informal protocol for display window controllers, which we had somehow missed in 8+ revisions of the 64-bit patch. With this declared as returning float, it caused subclass implementations?even those already correctly typed as returning CGFloat?to always return zero when running under 64-bit.